smart-waste-system/requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
opencv-python==4.9.0.80